postgreql connections much less forgiving on khan: processorInstance.cleanup() needed to be called in very many places. Similar fixes in a couple of other places, a nicer log message about dropping priority jobs if there are none, and probably a couple of other tweaks
git-svn-id: https://socorro.googlecode.com/svn/trunk@1733 efcc4bdc-162a-0410-a8af-5b4b46c5376d

+++ socorro/unittest/testlib/testDB.py
@@ -44,9 +44,12 @@ class TestDB:
   def removePriorityTables(self,config,logger):
     dbCon,dbCur = db_schema.connectToDatabase(config,logger)
     priorityTableNames = db_postgresql.tablesMatchingPattern('priority_job_%%',dbCur)
-    sql = "DROP TABLE IF EXISTS %s CASCADE;"%(", ".join(priorityTableNames))
-    dbCur.execute(sql)
-    dbCon.commit()
+    if priorityTableNames:
+      sql = "DROP TABLE IF EXISTS %s CASCADE;"%(", ".join(priorityTableNames))
+      dbCur.execute(sql)
+      dbCon.commit()
+    else:
+      logger.info("There were no priority_job tables to close")
     
   def populateDB(self,**kwargs):
     cursor = self.getCursor(**kwargs)
